+++ src/CoreModule.php
@@ -2,11 +2,12 @@
 
 namespace LaravelUi5\Core;
 
+use LaravelUi5\Core\Ui5\Contracts\Ui5Infrastructure;
 use LaravelUi5\Core\Ui5\Contracts\Ui5LibraryInterface;
 use LaravelUi5\Core\Ui5\AbstractUi5Module;
 
 
-class CoreModule extends AbstractUi5Module
+class CoreModule extends AbstractUi5Module implements Ui5Infrastructure
 {
     public function getName(): string
     {